requirements.txt
Django==1.8
django-cors-headers==2.1.0
django-rest-swagger==0.3.5
djangorestframework==3.3.3
djangorestframework-jwt==1.11.0